seperating gpt pipeline from in house pipeline
